+++ packages/esm-patient-common-lib/src/index.ts
@@ -7,6 +7,7 @@ export * from './error-state';
 export * from './form-entry-interop';
 export * from './form-entry/form-entry';
 export * from './get-patient-uuid-from-url';
+export * from './lab-resource';
 export * from './launchStartVisitPrompt';
 export * from './nav-group/createDashboardGroup';
 export * from './nav-group/DashboardGroupExtension';
