+++ packages/esm-patient-common-lib/src/lab-resource/index.ts
@@ -1,0 +1,2 @@
+export * from './lab.resource';
+export * from './test-order.component';
