+++ packages/esm-patient-common-lib/src/lab-resource/test-order.component.tsx
@@ -1,6 +1,6 @@
 import React, { useMemo } from 'react';
 import styles from './test-order.scss';
-import { type Order } from '@openmrs/esm-patient-common-lib';
+import { type Order, useLabEncounter, useOrderConceptByUuid } from '@openmrs/esm-patient-common-lib';
 import {
   DataTable,
   DataTableSkeleton,
@@ -14,14 +14,13 @@ import {
   TableRow,
 } from '@carbon/react';
 import { useTranslation } from 'react-i18next';
-import { useLabEncounter, useOrderConceptByUuid } from '../lab-results/lab-results.resource';
 import { useLayoutType } from '@openmrs/esm-framework';
 
 interface TestOrderProps {
   testOrder: Order;
 }
 
-const TestOrder: React.FC<TestOrderProps> = ({ testOrder }) => {
+export const TestOrder: React.FC<TestOrderProps> = ({ testOrder }) => {
   const { t } = useTranslation();
   const isTablet = useLayoutType() === 'tablet';
   const { concept, isLoading: isLoadingTestConcepts } = useOrderConceptByUuid(testOrder.concept.uuid);
@@ -112,5 +111,3 @@ const TestOrder: React.FC<TestOrderProps> = ({ testOrder }) => {
     </div>
   );
 };
-
-export default TestOrder;
