+++ packages/esm-patient-orders-app/translations/en.json
@@ -11,7 +11,6 @@
   "cancellationDateRequired": "Cancellation date is required",
   "cancelOrder": "Cancel order",
   "checkFilters": "Check the filters above",
-  "chooseAnOption": "Choose an option",
   "clearSearchResults": "Clear results",
   "dateCannotBeBeforeToday": "Date cannot be before today",
   "dateOfOrder": "Date of order",
@@ -23,28 +22,20 @@
   "dose": "Dose",
   "editResults": "Edit results",
   "endDate": "End date",
-  "enterTestResults": "Enter test results",
   "error": "Error",
   "errorCancellingOrder": "Error cancelling order",
   "errorFetchingTestTypes": "Error fetching results for \"{{searchTerm}}\"",
-  "errorSavingLabResults": "Error saving lab results",
   "goToDrugOrderForm": "Order form",
   "indication": "Indication",
-  "labResultError": "Error loading lab results",
   "inStock": "In stock",
   "instructions": "Instructions",
   "launchOrderBasket": "Launch order basket",
-  "loading": "Loading",
-  "loadingInitialValues": "Loading initial values",
-  "loadinglabresults": "Loading lab results",
-  "loadingTestDetails": "Loading test details",
   "medicationDurationAndUnit": "for {{duration}} {{durationUnit}}",
   "medicationIndefiniteDuration": "Indefinite duration",
   "medications": "Medications",
   "modifyOrder": "Modify order",
   "noMatchingOrdersToDisplay": "No matching orders to display",
   "noResultsForTestTypeSearch": "No results to display for \"{{searchTerm}}\"",
-  "normalRange": "Normal range",
   "onDate": "on",
   "or": "or",
   "order": "Order",
@@ -67,7 +58,6 @@
   "Orders": "Orders",
   "orderType": "Order type",
   "outOfStock": "Out of stock",
-  "pleaseFillField": "Please fill at least one field",
   "pleaseFillRequiredFields": "Please fill all the required fields",
   "pleaseRequiredFields": "Please fill all required fields",
   "price": "Price",
@@ -83,11 +73,9 @@
   "referenceNumberTableHeader": "{{orderType}} reference number",
   "refills": "Refills",
   "removeFromBasket": "Remove from basket",
-  "result": "Result",
   "returnToOrderBasket": "Return to order basket",
   "saveAndClose": "Save and close",
   "saveDrugOrderFailed": "Error ordering {{orderName}}",
-  "saveLabResults": "Save lab results",
   "saveOrder": "Save order",
   "saving": "Saving",
   "searchAgain": "search again",
@@ -102,7 +90,6 @@
   "startVisit": "Start visit",
   "status": "Status",
   "successfullyCancelledOrder": "Order {{orderNumber}} has been cancelled successfully",
-  "successfullySavedLabResults": "Lab results for {{orderNumber}} have been successfully updated",
   "Test Order_few": "Test orders",
   "testType": "Test type",
   "tryReopeningTheWorkspaceAgain": "Please try launching the workspace again",
