+++ packages/esm-patient-tests-app/src/lab-results/lab-result.scss
