+++ packages/esm-patient-tests-app/translations/en.json
@@ -7,6 +7,7 @@
   "backToTimeline": "Back to timeline",
   "cancel": "Cancel",
   "checkFilters": "Check the filters above",
+  "chooseAnOption": "chooseAnOption",
   "clear": "Clear",
   "clearFilters": "Clear filters",
   "clearSearchResults": "Clear results",
@@ -21,13 +22,20 @@
   "directlyAddToBasket": "Add to basket",
   "discard": "Discard",
   "endDate": "End date",
+  "enterTestResults": "enterTestResults",
   "error": "Error",
   "errorFetchingTestTypes": "Error fetching results for \"{{searchTerm}}\"",
+  "errorSavingLabResults": "errorSavingLabResults",
   "female": "Female",
   "goToDrugOrderForm": "Order form",
   "hideResultsTable": "Hide results table",
   "individualTests": "Individual tests",
   "labOrders": "Lab orders",
+  "labResultError": "labResultError",
+  "loading": "loading",
+  "loadingInitialValues": "loadingInitialValues",
+  "loadinglabresults": "loadinglabresults",
+  "loadingTestDetails": "loadingTestDetails",
   "male": "Male",
   "moreResultsAvailable": "More results available",
   "noPanelsFound": "No panels found",
@@ -48,6 +56,7 @@
   "overTime": "Over time",
   "panel": "Panel",
   "panels": "panels",
+  "pleaseFillField": "pleaseFillField",
   "pleaseRequiredFields": "Please fill all required fields",
   "print": "Print",
   "printedBy": "Printed by",
@@ -67,7 +76,10 @@
   "resultsText": "results",
   "returnToOrderBasket": "Return to order basket",
   "returnToTimeline": "Return to timeline",
+  "saveAndClose": "saveAndClose",
+  "saveLabResults": "saveLabResults",
   "saveOrder": "Save order",
+  "saving": "saving",
   "search": "Search",
   "searchAgain": "search again",
   "searchByTestName": "Search by test name",
@@ -80,6 +92,7 @@
   "showResultsTable": "Show results table",
   "showTree": "Show tree",
   "startDate": "Start date",
+  "successfullySavedLabResults": "successfullySavedLabResults",
   "testName": "Test name",
   "testResults": "test results",
   "testResults_title": "Test Results",
